Renames transformers2 and deletes dev files transformers_dev, transformers3
